Split large file into two parts
