+++ link-grammar/Makefile.am
@@ -107,6 +107,7 @@ liblink_grammar_la_SOURCES =        \
 	dict-atomese/read-atomese.c      \
 	dict-atomese/lookup-atomese.cc   \
 	dict-atomese/fetch-cats.cc       \
+	dict-atomese/sections.cc         \
 	dict-atomese/word-pairs.cc       \
 	disjunct-utils.c                 \
 	error.c                          \
